+++ ChangeLog.md
@@ -1,3 +1,12 @@
 # Changelog for TLT
 
+- 0.2.0 :: Divided the `tlt` function for running tests, to separate
+  `tltCore` for just running tests from formatted output of test
+  results.  The former is intended for running TLT in other
+  frameworks.
+
+- 0.1.0 :: First release.  Patch 1 re-arranged documentation.
+
 ## Unreleased changes
+
+None in this version.